fix: rename files, create directories, feat lockal pakage.json, fix root pakage.json, fix tsconfig
