+++ simple-auth/package.json
@@ -1,10 +1,14 @@
 {
-  "name": "template",
+  "name": "simple-auth",
+  "version": "0.0.1",
   "private": true,
-  "license": "BSD 3-Clause",
-  "workspaces": [
-    "simple-auth/**/*"
-  ],
+  "license": "BSD-3-Clause",
+  "main": "src/index.ts",
+  "scripts": {
+    "build": "yarn service build",
+    "dev": "yarn service dev",
+    "start": "node dist/index.js"
+  },
   "devDependencies": {
     "@atls/code-service": "^0.0.18",
     "@atls/config-jest": "^0.0.5",
@@ -17,6 +21,5 @@
     "eslint": "8.18.0",
     "husky": "^8.0.3",
     "typescript": "^4.6.3"
-  },
-  "packageManager": "yarn@3.5.1"
+  }
 }
