+++ simple-auth/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./simple-auth.interfaces";
+export * from "./login";
+export * from "./register";
+export * from "./whoami";
+export * from "./logout";
