+++ simple-auth/src/modules/support-modules/support-modules.check-auth-user-index.ts
@@ -1,0 +1,11 @@
+import { User } from "../../simple-auth.interfaces";
+
+export default async function checkAuthUserIndex(
+  usersList: User[]
+): Promise<number> {
+  const authUserIndex: number = await usersList.findIndex(
+    (user) => user.isAuth === true
+  );
+
+  return authUserIndex;
+}
