+++ simple-auth/src/register/register.hash-password.ts
