+++ simple-auth/src/simple-auth.interfaces.ts
