+++ simple-auth/src/support-modules/support-modules.check-auth-user-index.ts
@@ -1,4 +1,4 @@
-import { User } from "../../types";
+import { User } from "../simple-auth.interfaces";
 
 export default async function checkAuthUserIndex(
   usersList: User[]
