+++ simple-auth/src/support-modules/support-modules.constants.ts
